feat: use push notification hook
